Refactor unit tests
Unit tests should be in the same file as the code, to allow testing of
private functions.
