+++ src/main.rs
@@ -1,6 +1,3 @@
-#[cfg(test)]
-pub mod tests;
-
 pub mod game;
 pub mod game_logic;
 pub mod utils;
