--- src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod direction_tests;
-pub mod vector_tests;
